requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.5.7
charset-normalizer==3.1.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
lxml==4.9.2
multidict==6.0.4
numpy==1.24.3
openai==0.27.7
pandas==2.0.1
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tqdm==4.65.0
typing-extensions==4.6.0
tzdata==2023.3
urllib3==2.0.2
yarl==1.9.2
